refactor(chat): extract chat_manager/lifecycle_api.rs (#292)
Move the lifecycle and permission delegation methods off the facade
impl into their own submodule.
